state management preparation
