+++ lib/details_route.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class DetailsRoute extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Details'),
+      ),
+      body: Center(
+        child: Text("DETAILS PAGE"),
+      ),
+    );
+  }
+}
