+++ lib/state_providers_example_route.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class StateProvidersExampleRoute extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('State Example'),
+      ),
+      body: Center(
+        child: Text("State Management PAGE"),
+      ),
+    );
+  }
+}
